Add InsetLabel class
